test: finish modbus transport test decomposition
